requirements.txt
selenium-driverless==1.9.4
selenium==4.19.0
requests==2.31.0
pytz==2024.1
python-dotenv==1.0.1
pytest==8.1.1
ruff==0.11.13
black==23.9.1
boto3==1.35.82
Flask==3.1.0
ngrok==1.4.0
freezegun==1.5.1
pyautogui==0.9.54
google-api-python-client==2.165.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
discord.py==2.5.2
sqlalchemy==2.0.39
--only-binary psycopg2-binary
psycopg2-binary==2.9.9
alembic==1.15.2
